split things out
